+++ inc/packmodules.hpp
@@ -11,15 +11,15 @@ class TransferFcn: public PackModule {
 public:
     TransferFcn(Simulator *sim, const vecdble numerator, const vecdble denominator, std::string name="tf");
     virtual ~TransferFcn();
-    virtual PUnitModule Get_InputPort(uint n=0) const override;
-    virtual PUnitModule Get_OutputPort(uint n=0) const override;
+    virtual PUnitModule Get_InputPort(uint32_t n=0) const override;
+    virtual PUnitModule Get_OutputPort(uint32_t n=0) const override;
     void Set_InitialValue(vecdble value);
     vecdble Get_OutValue();
 private:
     UIntegrator **integrators = nullptr;
     USum *sum1 = nullptr;
     USum *sum2 = nullptr;
-    uint _order;
+    uint32_t _order;
 };
 
 
@@ -30,8 +30,8 @@ class DiscreteTransferFcn: public PackModule {
 public:
     DiscreteTransferFcn(Simulator *sim, const vecdble numerator, const vecdble denominator, std::string name="dtf");
     virtual ~DiscreteTransferFcn();
-    virtual PUnitModule Get_InputPort(uint n=0) const override;
-    virtual PUnitModule Get_OutputPort(uint n=0) const override;
+    virtual PUnitModule Get_InputPort(uint32_t n=0) const override;
+    virtual PUnitModule Get_OutputPort(uint32_t n=0) const override;
     void Set_SampleTime(double time);
     void Set_InitialValue(vecdble value);
     vecdble Get_OutValue();
@@ -39,7 +39,7 @@ private:
     UUnitDelay **unitdelays = nullptr;
     USum *sum1 = nullptr;
     USum *sum2 = nullptr;
-    uint _order;
+    uint32_t _order;
 };
 
 
@@ -50,8 +50,8 @@ class DiscreteIntegrator: public PackModule {
 public:
     DiscreteIntegrator(Simulator *sim, std::string name="disint");
     virtual ~DiscreteIntegrator() {}
-    virtual PUnitModule Get_InputPort(uint n=0) const override;
-    virtual PUnitModule Get_OutputPort(uint n=0) const override;
+    virtual PUnitModule Get_InputPort(uint32_t n=0) const override;
+    virtual PUnitModule Get_OutputPort(uint32_t n=0) const override;
     void Set_SampleTime(double time);
 private:
     UUnitDelay *delay1 = nullptr;
@@ -71,8 +71,8 @@ public:
     StateTransFcn(Simulator *sim, const zhnmat::Mat& A, const zhnmat::Mat& B, const zhnmat::Mat& C,
         bool isc=true, std::string name="stf");
     virtual ~StateTransFcn() {}
-    virtual PMatModule Get_InputBus(uint n=0) const override;
-    virtual PMatModule Get_OutputBus(uint n=0) const override;
+    virtual PMatModule Get_InputBus(uint32_t n=0) const override;
+    virtual PMatModule Get_OutputBus(uint32_t n=0) const override;
     void Set_InitialValue(const zhnmat::Mat& value);
     zhnmat::Mat Get_OutValue() const;
     void Set_SampleTime(double time);